WebSockets/tests/submissions/Opera/keeping-connection-open/001.html

<!doctype html>
<title>WebSockets: 20s inactivity after handshake</title>
<p>Note: This test takes 20 seconds to run.</p>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>
test(function(){
  var ws = new WebSocket(SCHEME_AND_DOMAIN+':8007/echo');
  ws.onclose = ws.onerror = ws.onmessage = assertUnreached;
  ws.onopen = function(e) {
    debug(e);
    setTestTimeout(30000);
    setTimeout(function() {
      ws.send('test');
      ws.onmessage = function(e) {
        debug(e);
        assert_equals(e.data, 'test');
        ws.onclose = function(e) {
          debug(e);
          setTestTimeout(null);
          setTimeout(end, 50);
        }
        ws.close();
      }
    }, 20000);
  }
});
</script>
